+++ sinp_nomenclatures/serializers.py
@@ -11,11 +11,13 @@ class NomenclatureSerializer(ModelSerializer):
 
     class Meta:
         model = Nomenclature
-        fields = ["id", "code", "label", "type", "status"]
+        fields = ["id", "code", "label", "type", "status", "parents"]
 
 
 class NomenclatureWithParentsSerializer(ModelSerializer):
     """Nomenclature item serializer"""
+
+    parents = NomenclatureSerializer(many=True)
 
     class Meta:
         model = Nomenclature